requirements.txt
Django==1.9.4
djangorestframework==3.3.2
nose==1.3.7
mysqlclient
django-mysql
django_extensions
django-cors-headers==2.0.2